more config options, changed spawning mechanics, fixed nbt parsing error (not cleanest but will work for now)
added more status effects when breaking a crystal
I got inspired by another mod/datapack called End evolved, where it gave players gravity to punish those who are flying.
I change it to instead of looking for flying players to just apply it to everyone present.
also since we dont have the gravity attribute yet, the work around was negative levitation via overflow (127+ levitation level).
I also changed how spawning of mobs work, you can specify in the config on the location to which they spawn, its not the best implimentation, but it works for now.
Also added more config options to choose from scaling or hard-coded limit.
umm the status effect numbers are pretty high, because this was for a modpack.